create many to many relationship between profile and posts. Website is no longer broken
